+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added Alarm plugin to Clock and Weather Station
 - KFC OTA copies firmware.elf and stores the hash in the device configuration
 - Rewritten StreamWrapper (Serial, DebugSerial, SerialHandler etc...)
+- Rewritten SerialHandler
 
 ## Version 0.0.3